+++ src/__tests__/changemakers.int.test.ts
@@ -4,7 +4,7 @@ import {
 	db,
 	createApplicationForm,
 	createApplicationFormField,
-	createBaseField,
+	createOrUpdateBaseField,
 	createOpportunity,
 	createChangemaker,
 	createChangemakerProposal,
@@ -344,7 +344,7 @@ describe('/changemakers', () => {
 				systemSource = await loadSystemSource(db, null);
 				systemUser = await loadSystemUser(db, null);
 				systemUserAuthContext = getAuthContext(systemUser);
-				baseFieldEmail = await createBaseField(db, null, {
+				baseFieldEmail = await createOrUpdateBaseField(db, null, {
 					label: 'Fifty one fifty three',
 					shortCode: 'fifty_one_fifty_three',
 					description: 'Five thousand one hundred fifty three.',
@@ -352,7 +352,7 @@ describe('/changemakers', () => {
 					scope: BaseFieldScope.ORGANIZATION,
 					valueRelevanceHours: null,
 				});
-				baseFieldPhone = await createBaseField(db, null, {
+				baseFieldPhone = await createOrUpdateBaseField(db, null, {
 					label: 'Fifty three ninety nine',
 					shortCode: 'fifty_three_ninety_nine',
 					description: 'Five thousand three hundred ninety nine.',
@@ -360,7 +360,7 @@ describe('/changemakers', () => {
 					scope: BaseFieldScope.ORGANIZATION,
 					valueRelevanceHours: null,
 				});
-				baseFieldWebsite = await createBaseField(db, null, {
+				baseFieldWebsite = await createOrUpdateBaseField(db, null, {
 					label: 'Fifty four seventy one 5471',
 					shortCode: 'fifty_four_seventy_one',
 					description: 'Five thousand four hundred seventy one.',
@@ -425,7 +425,7 @@ describe('/changemakers', () => {
 
 			it('returns the latest valid value for a base field when auth id is sent', async () => {
 				// Associate a base field associated with one opportunity/org, and add three responses.
-				const baseFieldId = baseFieldEmail.id;
+				const baseFieldShortCode = baseFieldEmail.shortCode;
 				const changemakerId = firstChangemaker.id;
 				const opportunityId = firstFunderOpportunity.id;
 				const proposalId = (
@@ -457,7 +457,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Org email',
 							applicationFormId: applicationFormIdEarliest,
-							baseFieldId,
+							baseFieldShortCode,
 							position: 5279,
 						})
 					).id,
@@ -483,7 +483,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Email contact',
 							applicationFormId: applicationFormIdLatestValid,
-							baseFieldId,
+							baseFieldShortCode,
 							position: 5347,
 						})
 					).id,
@@ -510,7 +510,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Contact email address',
 							applicationFormId: applicationFormIdLatest,
-							baseFieldId,
+							baseFieldShortCode,
 							position: 5209,
 						})
 					).id,
@@ -542,7 +542,7 @@ describe('/changemakers', () => {
 				const changemakerSourceId = secondChangemakerSourceId;
 				const funderSourceId = firstFunderSourceId;
 				// Associate one opportunity, one changemaker, and two responses with a base field.
-				const baseFieldId = baseFieldPhone.id;
+				const baseFieldShortCode = baseFieldPhone.shortCode;
 				const opportunity = firstFunderOpportunity;
 				const proposalId = (
 					await createProposal(db, systemUserAuthContext, {
@@ -575,7 +575,7 @@ describe('/changemakers', () => {
 							await createApplicationFormField(db, null, {
 								label: 'Org phone',
 								applicationFormId: applicationFormIdChangemakerEarliest,
-								baseFieldId,
+								baseFieldShortCode,
 								position: 5407,
 							})
 						).id,
@@ -603,7 +603,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Phone contact',
 							applicationFormId: applicationFormIdFunderLatest,
-							baseFieldId,
+							baseFieldShortCode,
 							position: 5417,
 						})
 					).id,
@@ -631,7 +631,7 @@ describe('/changemakers', () => {
 				const funderSourceId = firstFunderSourceId;
 				const dataProviderSourceId = firstDataProviderSourceId;
 				// Associate one opportunity, one changemaker, and two responses with a base field.
-				const baseFieldId = baseFieldPhone.id;
+				const baseFieldShortCode = baseFieldPhone.shortCode;
 				const opportunity = firstFunderOpportunity;
 				const proposalId = (
 					await createProposal(db, systemUserAuthContext, {
@@ -661,7 +661,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Organization phone 5437',
 							applicationFormId: applicationFormIdFunderEarliest,
-							baseFieldId,
+							baseFieldShortCode,
 							position: 5437,
 						})
 					).id,
@@ -688,7 +688,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Phone contact',
 							applicationFormId: applicationFormIdDataProviderLatest,
-							baseFieldId,
+							baseFieldShortCode,
 							position: 5443,
 						})
 					).id,
@@ -742,7 +742,7 @@ describe('/changemakers', () => {
 						await createApplicationFormField(db, null, {
 							label: 'Organization website 5479',
 							applicationFormId: applicationFormIdDataProviderEarliest,
-							baseFieldId: baseFieldWebsite.id,
+							baseFieldShortCode: baseFieldWebsite.shortCode,
 							position: 5479,
 						})
 					).id,
@@ -772,7 +772,7 @@ describe('/changemakers', () => {
 							await createApplicationFormField(db, null, {
 								label: 'Phone contact',
 								applicationFormId: applicationFormIdDataProviderLatest,
-								baseFieldId: baseFieldWebsite.id,
+								baseFieldShortCode: baseFieldWebsite.shortCode,
 								position: 5501,
 							})
 						).id,