+++ src/database/operations/baseFieldLocalization/loadBaseFieldLocalizationsBundleByBaseFieldShortCode.ts
@@ -1,0 +1,12 @@
+import { generateLoadBundleOperation } from '../generators';
+import type { BaseFieldLocalization, ShortCode } from '../../../types';
+
+const loadBaseFieldLocalizationsBundleByBaseFieldShortCode =
+	generateLoadBundleOperation<
+		BaseFieldLocalization,
+		[baseFieldShortCode: ShortCode]
+	>('baseFieldLocalizations.selectWithPagination', 'base_field_localizations', [
+		'baseFieldShortCode',
+	]);
+
+export { loadBaseFieldLocalizationsBundleByBaseFieldShortCode };