+++ src/database/operations/baseFields/createOrUpdateBaseField.ts
@@ -1,12 +1,12 @@
 import { generateCreateOrUpdateItemOperation } from '../generators';
-import type { BaseField, WritableBaseField } from '../../../types';
+import type { BaseField, InternallyWritableBaseField } from '../../../types';
 
 const createOrUpdateBaseField = generateCreateOrUpdateItemOperation<
 	BaseField,
-	WritableBaseField,
+	InternallyWritableBaseField,
 	[]
 >(
-	'baseFields.createOrUpdateByShortcode',
+	'baseFields.createOrUpdateOne',
 	[
 		'scope',
 		'dataType',