+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ProposalFieldValues` now have a `goodAsOf` attribute.
 
+## 0.18.0 2025-04-28
+
+### Changed
+
+- `BaseFields` are now primary keyed on the `shortCode` value, and have had corresponding
+  foreign keys in the `ApplicationFormField` and `BaseFieldLocalization` tables updated to match.
+
 ## 0.17.1 2025-04-21
 
 ### Added